refactor: 리뷰내용 반영
패키지 일괄 이동
  - step3 -> edu.nextstep.racing

메소드 순서 조정
  - Game, Playgound 클래스

미사용 필드 제거
  - ResultView: scanner

테스트 순서 정의 제거 & 불필요한 반복테스트 제거
  - GameTest, NumberUtilTest
