+++ src/main/java/edu/nextstep/racing/NumberUtil.java
@@ -1,4 +1,4 @@
-package step3;
+package edu.nextstep.racing;
 
 import java.util.Random;
