+++ src/main/java/edu/nextstep/racing/Playground.java
@@ -1,9 +1,20 @@
-package step3;
+package edu.nextstep.racing;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class Playground {
+
+	public void play() {
+		int numberOfCars = InputView.inputNumberOfCars();
+		int moveTryMaxCount = InputView.inputMoveTryMaxCount();
+		Game racingGame = createRacingGame(numberOfCars, moveTryMaxCount);
+		ResultView.printGameStart();
+		while (racingGame.isContinue()) {
+			racingGame.play();
+			ResultView.printCarsLocation(racingGame.getGameCars());
+		}
+	}
 
 	public Game createRacingGame(int numberOfCars, int moveTryMaxCount) {
 		List<Car> gameCars = createGameCars(numberOfCars);
@@ -16,16 +27,5 @@ public class Playground {
 			entry.add(new Car());
 		}
 		return entry;
-	}
-
-	public void play() {
-		int numberOfCars = InputView.inputNumberOfCars();
-		int moveTryMaxCount = InputView.inputMoveTryMaxCount();
-		Game racingGame = createRacingGame(numberOfCars, moveTryMaxCount);
-		ResultView.printGameStart();
-		while (racingGame.isContinue()) {
-			racingGame.play();
-			ResultView.printCarsLocation(racingGame.getGameCars());
-		}
 	}
 }
