+++ src/test/java/edu/nextstep/racing/GameTest.java
@@ -1,4 +1,4 @@
-package step3;
+package edu.nextstep.racing;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -6,16 +6,11 @@ import java.util.Arrays;
 import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.MethodOrderer;
-import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
 
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @DisplayName("Game : 자동차 게임을 진행함에 필요한 상태들을 가지고 있고, 게임을 진행하는 등의 역할을 수행함.")
 class GameTest {
 
-	@Order(1)
 	@DisplayName("getGameCars : 자동차 게임에 참가한 자동차들 목록을 가져옴.")
 	@Test
 	void getGameCars() {
@@ -24,7 +19,6 @@ class GameTest {
 		assertThat(new Game(gameCars, 5).getGameCars()).isEqualTo(gameCars);
 	}
 
-	@Order(2)
 	@DisplayName("getMoveTryMaxCount : 자동차 게임에서 총 몇 번의 이동을 할 것인지 횟수를 구함.")
 	@Test
 	void getMoveTryMaxCount() {
@@ -33,7 +27,6 @@ class GameTest {
 		assertThat(new Game(entry, 5).getMoveTryMaxCount()).isEqualTo(5);
 	}
 
-	@Order(3)
 	@DisplayName("tryToMoveAllGameCars : 자동차 게임에 참가한 모든 자동차들의 위치가 1 또는 0 증가하여야 함.")
 	@Test
 	void tryToMoveAllGameCars() {
@@ -48,7 +41,6 @@ class GameTest {
 		);
 	}
 
-	@Order(4)
 	@DisplayName("play : 자동차 게임을 1회 진행함. 참가한 모든 자동차들은 위치가 0 이상 누적 이동횟수 이하여여 함.")
 	@Test
 	void play() {
@@ -63,7 +55,6 @@ class GameTest {
 		);
 	}
 
-	@Order(5)
 	@DisplayName("isContinue : 현재까지 누적 이동 횟수가 Game에 사전 설정된 최대 이동횟수를 넘지 않으면 true를 리턴함.")
 	@Test
 	void isContinue() {
@@ -78,7 +69,6 @@ class GameTest {
 		assertThat(game.isContinue()).isFalse();
 	}
 
-	@Order(6)
 	@DisplayName("play와 isContinue : 자동차 게임을 끝까지 진행함. 참가한 모든 자동차들은 위치가 0 이상 최대 이동 시도 횟수 이하여여 함.")
 	@Test
 	void playToEnd() {
